+++ README.md
@@ -16,56 +16,16 @@ pip install git+https://github.com/Synternet/pubsub-python.git
 
 Before you begin using the Python SDK, make sure you have the necessary credentials and access tokens from the Synternet [Developer Portal](https://portal.synternet.com/) platform. These credentials will allow you to connect to the Data Layer and subscribe to or publish data streams.
 
+More information available in [https://docs.synternet.com/docs](Synternet documentation).
+
 ## Examples
 
 For detailed usage examples, please refer to the [examples directory](https://github.com/Synternet/pubsub-python/examples) in the repository. These examples cover various scenarios and demonstrate how to utilize the SDK's features effectively.
 
 The preferred authentication method is using an access token from the [developer portal](https://portal.synternet.com/).
 
-```Text Python
-import asyncio
-import nats
-import tempfile
-import os
-
-from helper import create_app_jwt
-
-ACCESS_TOKEN = "EXAMPLE_ACCESS_TOKEN"
-
-async def main():
-    jwt = create_app_jwt(ACCESS_TOKEN)
-
-    # Write the JWT to a temporary file with correct format
-    with tempfile.NamedTemporaryFile(mode='w+t', delete=False) as temp:
-        temp.write(jwt)
-        temp_path = temp.name
-
-    nc = await nats.connect("nats://127.0.0.1", user_credentials=temp_path)
-
-    async def message_handler(msg):
-        subject = msg.subject
-        data = msg.data.decode()
-        print("Received a message on '{subject}: {data}".format(
-            subject=subject, data=data))
-
-    await nc.subscribe("syntropy.bitcoin.tx", cb=message_handler)
-
-    # run infinitely
-    while True:
-        await asyncio.sleep(1)
-
-    # Terminate connection to NATS.
-    await nc.drain()
-
-    # Delete the temporary file
-    os.unlink(temp_path)
-
-if __name__ == '__main__':
-    asyncio.run(main())
-```
-
 # Contributing
-We welcome contributions from the community! If you find any issues or have suggestions for improvements, please open an issue or submit a pull request on the [GitHub repository](https://github.com/Synternet/pubsub-python). We appreciate your feedback and collaboration in making this SDK even better. 
+We welcome contributions from the community! If you find any issues or have suggestions for improvements, please open an issue or submit a pull request on the [GitHub repository](https://github.com/Synternet/pubsub-python). We appreciate your feedback and collaboration in making this SDK even better.
 
 ## Contribution Guidelines
 
@@ -81,7 +41,6 @@ When making changes to the codebase, it's important to follow a consistent commi
 To streamline the pull request process, we have provided a [Pull Request Template](pull-request-template.md) that includes the necessary sections for describing your changes, related issues, proposed changes, and any additional information. Make sure to fill out the template when submitting a pull request.
 
 We appreciate your contributions and thank you for your support in making this project better!
-
 
 # Support
