+++ examples/republish_using_creds_files.py
@@ -1,7 +1,10 @@
 import asyncio
 import nats
 
-NATS_URL = "nats://127.0.0.1"
+# See https://docs.synternet.com/build/dl-access-points
+NATS_URL = "nats://broker-eu-01.synternet.com"
+SUB_SUBJECT = "synternet.example.subject"
+PUB_SUBJECT = "republisher.example.subject"
 
 async def main():
     sub_nc = await nats.connect(NATS_URL, user_credentials="subscriber_nats.creds")
@@ -12,9 +15,9 @@ async def main():
         data = msg.data.decode()
         print("Received a message on '{subject}: {data}".format(
             subject=subject, data=data))
-        await pub_nc.publish("republisher.example.subject", msg.data)
+        await pub_nc.publish(PUB_SUBJECT, msg.data)
 
-    await sub_nc.subscribe("synternet.example.subject", cb=message_handler)
+    await sub_nc.subscribe(SUB_SUBJECT, cb=message_handler)
 
     # run infinitely
     while True:
