+++ examples/subscribe_with_seed.py
@@ -5,7 +5,10 @@ import nats
 
 from helper import create_app_jwt
 
-ACCESS_TOKEN = "EXAMPLE_ACCESS_TOKEN"
+# See https://docs.synternet.com/build/dl-access-points
+NATS_URL = "nats://broker-eu-01.synternet.com"
+ACCESS_TOKEN = "ACCESS_TOKEN"
+STREAM_SUBJECT = "synternet.example.subject"
 
 async def main():
     jwt = create_app_jwt(ACCESS_TOKEN)
@@ -15,7 +18,7 @@ async def main():
         temp.write(jwt)
         temp_path = temp.name
 
-    nc = await nats.connect("nats://127.0.0.1", user_credentials=temp_path)
+    nc = await nats.connect(NATS_URL, user_credentials=temp_path)
 
     async def message_handler(msg):
         subject = msg.subject
@@ -23,7 +26,7 @@ async def main():
         print("Received a message on '{subject}: {data}".format(
             subject=subject, data=data))
 
-    await nc.subscribe("synternet.example.subject", cb=message_handler)
+    await nc.subscribe(STREAM_SUBJECT, cb=message_handler)
 
     # run infinitely
     while True:
